+++ src/tx-data/create-swap-pool.js
@@ -1,0 +1,39 @@
+import defineProperties from '../define-properties.js';
+
+export default class TxDataCreateSwapPool {
+    constructor(data) {
+        // Define Properties
+        const fields = [
+            {
+                name: 'coin0',
+                length: 4,
+                allowLess: true,
+            },
+            {
+                name: 'coin1',
+                length: 4,
+                allowLess: true,
+            },
+            {
+                name: 'volume0',
+                length: 32,
+                allowLess: true,
+            },
+            {
+                name: 'volume1',
+                length: 32,
+                allowLess: true,
+            },
+        ];
+
+        /**
+         * Returns the rlp encoding of the transaction
+         * @method serialize
+         * @return {Buffer}
+         * @memberof Transaction
+         * @name serialize
+         */
+        // attached serialize
+        defineProperties(this, fields, data);
+    }
+}