+++ test/tx-data/vote-halt-block.test.js
@@ -1,0 +1,23 @@
+import {TX_TYPE, TxData, TxDataSetHaltBlock} from '~/src';
+import decodeToArray from '../decode-to-array';
+
+describe('TxDataVoteHaltBlock', () => {
+    const params = {
+        publicKey: 'Mpf9e036839a29f7fba2d5394bd489eda927ccb95acc99e506e688e4888082b3a3',
+        height: 1000,
+    };
+
+    const serializedTxData = (new TxDataSetHaltBlock(params)).serialize();
+
+    test('rlp encoded fields', () => {
+        expect(decodeToArray(serializedTxData))
+            .toEqual([
+                [249, 224, 54, 131, 154, 41, 247, 251, 162, 213, 57, 75, 212, 137, 237, 169, 39, 204, 185, 90, 204, 153, 229, 6, 230, 136, 228, 136, 128, 130, 179, 163],
+                [3, 232],
+            ]);
+    });
+
+    test('TxData', () => {
+        expect(TxData(params, TX_TYPE.SET_HALT_BLOCK).serialize()).toEqual(serializedTxData);
+    });
+});